FEATURE: Improving thread list item and header
* Moved the settings cog from thread list to thread and
  put it in a new header element
* Remove thread original message component, no longer needed
  and the list item and thread indicator styles/content
  will be quite different
* Start adding content (unread indicator etc.) to the thread
  list item
