+++ plugins/chat/assets/javascripts/discourse/components/chat-thread.hbs
@@ -9,17 +9,7 @@
   {{will-destroy this.unsubscribeFromUpdates}}
 >
   {{#if @includeHeader}}
-    <div class="chat-thread__header">
-      <span class="chat-thread__label">{{i18n "chat.thread.label"}}</span>
-      <LinkTo
-        class="chat-thread__close btn-flat btn btn-icon no-text"
-        @route="chat.channel"
-        @models={{this.channel.routeModels}}
-        title={{i18n "chat.thread.close"}}
-      >
-        {{d-icon "times"}}
-      </LinkTo>
-    </div>
+    <Chat::Thread::Header @thread={{this.thread}} @channel={{this.channel}} />
   {{/if}}
 
   <div
