+++ plugins/chat/assets/javascripts/discourse/components/chat/thread/header-unread-indicator.js
@@ -1,9 +1,6 @@
 import Component from "@glimmer/component";
-import { inject as service } from "@ember/service";
 
 export default class ChatThreadHeaderUnreadIndicator extends Component {
-  @service chatTrackingStateManager;
-
   get showUnreadIndicator() {
     return this.args.channel.unreadThreadCount > 0;
   }
