+++ plugins/chat/assets/javascripts/discourse/components/chat/thread/header.hbs
@@ -1,0 +1,25 @@
+<div class="chat-thread-header">
+  <span class="chat-thread-header__label overflow-ellipsis">
+    {{replace-emoji this.label}}
+  </span>
+
+  <div class="chat-thread-header__buttons">
+    {{#if this.canChangeThreadSettings}}
+      <DButton
+        @action={{action this.openThreadSettings}}
+        @class="btn-flat chat-thread-list-item__settings"
+        @icon="cog"
+        @title="chat.thread.settings"
+      />
+    {{/if}}
+
+    <LinkTo
+      class="chat-thread__close btn-flat btn btn-icon no-text"
+      @route="chat.channel"
+      @models={{@channel.routeModels}}
+      title={{i18n "chat.thread.close"}}
+    >
+      {{d-icon "times"}}
+    </LinkTo>
+  </div>
+</div>
