+++ plugins/chat/assets/javascripts/discourse/components/chat/thread/list-item-unread-indicator.hbs
@@ -1,0 +1,9 @@
+{{#if this.showUnreadIndicator}}
+  <div class="chat-thread-list-item-unread-indicator">
+    <div class="chat-thread-list-item-unread-indicator__number-wrap">
+      <div
+        class="chat-thread-list-item-unread-indicator__number"
+      >{{this.unreadCountLabel}}</div>
+    </div>
+  </div>
+{{/if}}
