+++ plugins/chat/assets/javascripts/discourse/components/chat/thread/list.hbs
@@ -5,17 +5,7 @@
   {{will-destroy this.teardown}}
 >
   {{#if @includeHeader}}
-    <div class="chat-thread__header">
-      <span class="chat-thread__label">{{i18n "chat.threads.list"}}</span>
-      <LinkTo
-        class="chat-thread__close btn-flat btn btn-icon no-text"
-        @route="chat.channel"
-        @models={{@channel.routeModels}}
-        title={{i18n "chat.thread.close"}}
-      >
-        {{d-icon "times"}}
-      </LinkTo>
-    </div>
+    <Chat::Thread::Header @channel={{@channel}} />
   {{/if}}
 
   <div class="chat-thread-list__items">
