--- plugins/chat/assets/javascripts/discourse/components/chat/thread/original-message.hbs
@@ -1,15 +1,0 @@
-<div class="chat-thread-original-message">
-  <div class="chat-thread-original-message__inner-container">
-    <div class="chat-thread-original-message__excerpt">
-      {{replace-emoji (html-safe @message.excerpt)}}
-    </div>
-    <div class="chat-thread-original-message__author">
-      <span class="chat-thread-original-message__avatar">
-        <ChatUserAvatar @user={{@message.user}} />
-      </span>
-      <span class="chat-thread-original-message__username">
-        {{@message.user.username}}
-      </span>
-    </div>
-  </div>
-</div>
