--- plugins/chat/assets/javascripts/discourse/components/chat/thread/original-message.js
@@ -1,3 +1,0 @@
-import Component from "@glimmer/component";
-
-export default class ChatThreadOriginalMessage extends Component {}
