--- plugins/chat/assets/javascripts/discourse/components/styleguide/chat-thread-original-message.hbs
@@ -1,5 +1,0 @@
-<StyleguideExample @title="<ChatThreadOriginalMessage>">
-  <Styleguide::Component>
-    <Chat::Thread::OriginalMessage @message={{this.message}} />
-  </Styleguide::Component>
-</StyleguideExample>
