+++ plugins/chat/assets/stylesheets/common/chat-thread-header.scss
@@ -1,0 +1,15 @@
+.chat-thread-header {
+  height: var(--chat-header-offset);
+  min-height: var(--chat-header-offset);
+  border-bottom: 1px solid var(--primary-low);
+  border-top: 1px solid var(--primary-low);
+  box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding-inline: 1rem;
+
+  &__buttons {
+    display: flex;
+  }
+}
