+++ plugins/chat/assets/stylesheets/common/chat-thread-list-item.scss
@@ -10,6 +10,11 @@
 .chat-thread-list-item {
   @include thread-list-item;
   cursor: pointer;
+  margin: 0.75rem 0.25rem 0.75rem 0.5rem;
+
+  & + .chat-thread-list-item {
+    margin-top: 0;
+  }
 
   .touch & {
     &:active {
@@ -28,21 +33,33 @@
 
   &__main {
     flex: 1 1 100%;
+    width: 100%;
+  }
+
+  &__body {
+    padding-bottom: 0.25rem;
+    word-break: break-word;
+    margin: 0.5rem 0rem;
+
+    > * {
+      pointer-events: none;
+    }
   }
 
   &__header {
     display: flex;
     flex-direction: row;
     align-items: center;
+    margin-bottom: 0.25rem;
   }
 
   &__title {
-    flex: 1;
+    flex: 1 1 auto;
     font-weight: bold;
   }
 
-  &__settings[disabled] {
-    display: none;
+  &__unread-indicator {
+    flex: 0 0 auto;
   }
 
   &__open-button {
@@ -56,5 +73,10 @@
     &:visited {
       color: var(--primary);
     }
+  }
+
+  &__om-user-avatar {
+    margin-right: 0.5rem;
+    flex: 0 0 auto;
   }
 }
