--- plugins/chat/assets/stylesheets/common/chat-thread-original-message.scss
@@ -1,26 +1,0 @@
-.chat-thread-original-message {
-  display: flex;
-  margin: 0.5rem 0rem;
-
-  &__inner-container {
-    width: 100%;
-  }
-
-  &__excerpt {
-    padding-bottom: 0.25rem;
-    word-break: break-word;
-
-    > * {
-      pointer-events: none;
-    }
-  }
-
-  &__author {
-    display: flex;
-    align-items: center;
-  }
-
-  &__avatar {
-    padding: 0.25rem 0.25rem 0.25rem 0;
-  }
-}
