+++ plugins/chat/assets/stylesheets/common/chat-thread-unread-indicator.scss
@@ -1,0 +1,26 @@
+@mixin chat-thread-unread-indicator {
+  color: var(--tertiary);
+  padding-left: 0.25rem;
+
+  &__number-wrap {
+    background-color: var(--tertiary-med-or-tertiary);
+    display: flex;
+    padding: 0.25rem 0.5rem;
+    border-radius: 20px;
+    width: 35px;
+    box-sizing: border-box;
+    flex-direction: column;
+    align-items: center;
+  }
+
+  &__number {
+    color: var(--secondary);
+    font-size: var(--font-down-3);
+    font-weight: bold;
+  }
+}
+
+.chat-thread-header-unread-indicator,
+.chat-thread-list-item-unread-indicator {
+  @include chat-thread-unread-indicator;
+}
