+++ plugins/chat/assets/stylesheets/common/chat-thread.scss
@@ -4,18 +4,6 @@
   position: relative;
   @include chat-height;
 
-  &__header {
-    height: var(--chat-header-offset);
-    min-height: var(--chat-header-offset);
-    border-bottom: 1px solid var(--primary-low);
-    border-top: 1px solid var(--primary-low);
-    box-sizing: border-box;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding-inline: 1rem;
-  }
-
   &__body {
     overflow-y: scroll;
     @include chat-scrollbar();
