+++ plugins/chat/assets/stylesheets/common/chat-threads-list.scss
@@ -12,18 +12,6 @@
     overscroll-behavior: contain;
     display: flex;
     flex-direction: column;
-
-    .chat-thread-list-item {
-      margin: 0.75rem 0.25rem 0.75rem 0.5rem;
-
-      &.-unread {
-        border-left: 2px solid var(--tertiary-medium);
-      }
-
-      & + .chat-thread-list-item {
-        margin-top: 0;
-      }
-    }
   }
 
   &__no-threads {
