+++ plugins/chat/assets/stylesheets/common/index.scss
@@ -53,3 +53,5 @@
 @import "chat-thread-original-message";
 @import "chat-composer-separator";
 @import "chat-thread-header-button";
+@import "chat-thread-header";
+@import "chat-thread-unread-indicator";
